chore: Refactor property definitions logic to separate class
