+++ plugin-server/src/worker/ingestion/process-event.ts
@@ -27,6 +27,7 @@ import { GroupTypeManager } from './group-type-manager'
 import { addGroupProperties } from './groups'
 import { LazyPersonContainer } from './lazy-person-container'
 import { upsertGroup } from './properties-updater'
+import { PropertyDefinitionsManager } from './property-definitions-manager'
 import { TeamManager } from './team-manager'
 import { captureIngestionWarning } from './utils'
 
@@ -37,6 +38,7 @@ export class EventsProcessor {
     kafkaProducer: KafkaProducerWrapper
     teamManager: TeamManager
     groupTypeManager: GroupTypeManager
+    propertyDefinitionsManager: PropertyDefinitionsManager
 
     constructor(pluginsServer: Hub) {
         this.pluginsServer = pluginsServer
@@ -45,6 +47,12 @@ export class EventsProcessor {
         this.kafkaProducer = pluginsServer.kafkaProducer
         this.teamManager = pluginsServer.teamManager
         this.groupTypeManager = new GroupTypeManager(pluginsServer.db, this.teamManager, pluginsServer.SITE_URL)
+        this.propertyDefinitionsManager = new PropertyDefinitionsManager(
+            this.teamManager,
+            pluginsServer.db,
+            pluginsServer,
+            pluginsServer.statsd
+        )
     }
 
     public async processEvent(
@@ -156,7 +164,7 @@ export class EventsProcessor {
             properties['$ip'] = ip
         }
 
-        await this.teamManager.updateEventNamesAndProperties(team.id, event, properties)
+        await this.propertyDefinitionsManager.updateEventNamesAndProperties(team.id, event, properties)
         properties = await addGroupProperties(team.id, properties, this.groupTypeManager)
 
         if (event === '$groupidentify') {
